pyproject.toml
[project]
name = "crawl4ai-mcp"
version = "0.1.0"
description = "Agentic MCP server for intelligent web crawling and RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai==0.6.2",
    "mcp==1.7.1",
    "supabase==2.15.1",
    "openai==1.71.0",
    "anthropic>=0.52.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "playwright>=1.40.0"
]